+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 5.08.2018
+
+* [Rust] Code of Kutulu - Wanderer state enum (CodinGame)
+
 #### 4.08.2018
 
 * [Rust] Code of Kutulu - Entity distance in use (CodinGame)